검색, 앨범목록 컴포넌트 분리
