+++ src/App.scss
@@ -3,27 +3,6 @@
 @import './assets/scss/common.scss';
 
 /* TODO: 각 컴포넌트 스타일로 위치 변경 */
-.search {
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  flex-wrap: nowrap;
-  padding: 15px;
-  background: lighten(gray, 40%);
-
-  input {
-    width: 100%;
-    height: 45px;
-    box-sizing: border-box;
-    margin: 0;
-    padding: 10px;
-  }
-
-  button {
-    width: 80px;
-    margin-left: 10px;
-  }
-}
 
 .album-item {
   display: flex;
