+++ src/assets/scss/reset.scss
@@ -9,3 +9,10 @@ ul, ol, li {
   padding: 0;
   list-style: none;
 }
+
+input {
+  margin: 0;
+  padding: 0;
+  background: 0 none;
+  border: 0 none;
+}
