+++ src/components/Search/style.scss
@@ -1,0 +1,21 @@
+.search-form {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  flex-wrap: nowrap;
+  padding: 15px;
+  background: lighten(gray, 40%);
+
+  input {
+    width: 100%;
+    height: 45px;
+    box-sizing: border-box;
+    margin: 0;
+    padding: 10px;
+  }
+
+  button {
+    width: 80px;
+    margin-left: 10px;
+  }
+}
